+++ packages/gov-ui-kit/src/components/dialogs/dialogAlert/dialogAlertContent/index.ts
@@ -1,0 +1,1 @@
+export { DialogAlertContent, type IDialogAlertContentProps } from './dialogAlertContent';