+++ packages/gov-ui-kit/src/components/dialogs/dialogAlert/dialogAlertRoot/index.ts
@@ -1,0 +1,7 @@
+export {
+    DialogAlertContext,
+    DialogAlertRoot,
+    type DialogAlertVariant,
+    type IDialogAlertContext,
+    type IDialogAlertRootProps,
+} from './dialogAlertRoot';