+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `DialogAlert` component
+
+### Fixed
+
+-   `Icon` styling to prevent shrinking
+
 ## [1.0.15] - 2024-02-23
 
 ### Changed